v1.5.8 -Changed Fault to use logger and moved enums in fault to new condition package in util
