+++ TeamCode/src/main/java/util/codeseg/BooleanCodeSeg.java
@@ -1,4 +1,4 @@
-package util.CodeSegs;
+package util.codeseg;
 
 public interface BooleanCodeSeg {
     boolean run();
