+++ TeamCode/src/main/java/util/codeseg/CodeSeg.java
@@ -1,4 +1,4 @@
-package util.CodeSegs;
+package util.codeseg;
 
 //Used to define and pass code
 public interface CodeSeg {
