+++ TeamCode/src/main/java/util/codeseg/DoubleCodeSeg.java
@@ -1,4 +1,4 @@
-package util.CodeSegs;
+package util.codeseg;
 
 public interface DoubleCodeSeg {
     double run();
