+++ TeamCode/src/main/java/util/condition/Expectation.java
@@ -1,0 +1,10 @@
+package util.condition;
+
+public enum Expectation{
+    TRIVIAL,
+    OBVIOUS,
+    EXPECTED,
+    SURPRISING,
+    UNEXPECTED,
+    INCONCEIVABLE
+}
