+++ TeamCode/src/main/java/util/condition/Magnitude.java
@@ -1,0 +1,10 @@
+package util.condition;
+
+public enum Magnitude{
+    NEGLIGIBLE,
+    MINOR,
+    MODERATE,
+    MAJOR,
+    CRITICAL,
+    CATASTROPHIC
+}
